Refactored stored file remote upload to use a service
